requirement.txt
streamlit
google-generativeai
python-dotenv
langchain-community
PyPDF2
faiss-cpu 
langchain_google_genai